+++ chatlas/__init__.py
@@ -13,7 +13,7 @@ from ._perplexity import ChatPerplexity
 from ._provider import Provider
 from ._snowflake import ChatSnowflake
 from ._tokens import token_usage
-from ._tools import Tool
+from ._tools import Tool, ToolResult
 from ._turn import Turn
 
 try:
@@ -43,6 +43,7 @@ __all__ = (
     "Provider",
     "token_usage",
     "Tool",
+    "ToolResult",
     "Turn",
     "types",
 )