+++ chatlas/_content.py
@@ -3,7 +3,15 @@ from __future__ import annotations
 import json
 from dataclasses import dataclass
 from pprint import pformat
-from typing import Any, Literal, Optional
+from typing import TYPE_CHECKING, Any, Literal, Optional, Protocol, cast
+
+if TYPE_CHECKING:
+    from ._tools import ToolResult
+
+
+class Stringable(Protocol):
+    def __str__(self) -> str: ...
+
 
 ImageContentTypes = Literal[
     "image/png",
@@ -195,20 +203,21 @@ class ContentToolResult(Content):
     """
 
     id: str
-    value: Any = None
+    result: Optional[ToolResult] = None
     name: Optional[str] = None
     error: Optional[str] = None
 
-    def _get_value(self, pretty: bool = False) -> str:
+    def _get_value(self, pretty: bool = False) -> Stringable:
         if self.error:
             return f"Tool calling failed with error: '{self.error}'"
+        result = cast("ToolResult", self.result)
         if not pretty:
-            return str(self.value)
+            return result.assistant
         try:
-            json_val = json.loads(self.value)  # type: ignore
+            json_val = json.loads(result.assistant)  # type: ignore
             return pformat(json_val, indent=2, sort_dicts=False)
         except:  # noqa: E722
-            return str(self.value)
+            return result.assistant
 
     # Primarily used for `echo="all"`...
     def __str__(self):
@@ -222,13 +231,14 @@ class ContentToolResult(Content):
 
     def __repr__(self, indent: int = 0):
         res = " " * indent
-        res += f"<ContentToolResult value='{self.value}' id='{self.id}'"
+        value = None if self.result is None else self.result.assistant
+        res += f"<ContentToolResult value='{value}' id='{self.id}'"
         if self.error:
             res += f" error='{self.error}'"
         return res + ">"
 
     # The actual value to send to the model
-    def get_final_value(self) -> str:
+    def get_final_value(self) -> Stringable:
         return self._get_value()